echem_app/requirements.txt
Flask >= 3.0.0
matplotlib >= 3.5.0
numpy >= 1.21.1
scipy >= 1.10.0
pyserial >= 3.5
bleak >= 0.21.0       # BLE — required for PS-539B over Bluetooth
zeroconf >= 0.130.0   # mDNS — publishes the app as neotrient.local on the LAN
qrcode >= 7.4.0       # QR code generation for the server-status page
# Optional — only loaded if you have a real model file
# torch >= 2.0.0
# joblib >= 1.3.0
# scikit-learn >= 1.3.0
